Feat: Setting up Docker Compose file for Deployment
Created a docker compose deploy file which consists of different services being used inside the application. We make sure to configure all the servies to automatically start if failed or stopped for any reason. Also we add in persistent storage volumes for services: 'app' and 'proxy'. To make sure, one service does not override another, we add in 'depends_on' section in the service.